fix coverity issues

+++ gazebo/math/Helpers.hh
@@ -240,7 +240,7 @@ namespace gazebo
         p++;
       }
 
-      double acc = 0;
+      unsigned int acc = 0;
       while (*p >= '0' && *p <= '9')
         acc = acc * 10 + *p++ - '0';
